requirements.txt
numpy
roboticstoolbox-python
sympy
matplotlib
uModbus
pyserial